Version 0.15
The enemy can now attack the player. I also coded a simple UI which contains health and mana bars - only the health bar works currently and will drop as the player takes damage from enemies. Once the player hits 0 health, a separate UI will show which has a death notification.

I still plan to make it so the enemy can hear the player when they sprint or walk, but I need to implement the ability for the player to sprint and crouch first.

Next up, I think I will code the mana system and then an arrow system which limits the arrows a player has.