Version 0.16
I implemented the mana system and linked it to the staff so that if a player attacks with the staff, they will lose some mana. If the player does not have enough mana to attack, then the staff will not create a projectile.

I also made it so that after mana is lost, a 2 second cooldown begins and then mana will start regenerating in increments of 10 every second. However, if the player spends mana during regeneration then they will have to wait another 2 seconds before regeneration will start again.

As I stated in the previous version, I plan to now work on implementing an ammo system for the bow (this should be a quick task as it should be similar to the staff/mana system).