Version 0.18
Started implementing the inventory system. Currently, you can press TAB to open and close the inventory. When open, the player cannot attack or look around and the cursor becomes unlocked/visible. This is reverted once the inventory is closed.

I also made a prefab for an inventory item - I will be using this once I make further developments with the inventory system.