Version 0.19
I have remade a bit of the inventory so that it is a panel and not a scroll view now. This is because I do not want the player to be able to scroll in their inventory.

I have also made some basic item interaction where the player can press E and if an interactable item is looked at and close enough, it will be deleted.

I will be further developing this so that the item will be added to the inventory.