Version 0.21
I worked on finishing the movement system - now the player can sprint and move faster or enter 'stealth' mode and move slower. The enemy can also detect whether the player is stealthing or not - if they are then the enemy will not be able to hear the player. However, if the player is walking or sprinting and moving at a certain speed, the enemy is able to hear the player and can target them without seeing them directly.