Version 0.25
I have added consumable item types - so far these are health potions, mana potions and arrows. These are items which, when used in the inventory, change the player's stats.

I also added some more equipment and finished implementing the modifiers. Now, incoming damage is negated by the armour value - this applies for damage received for both the player and enemies (if an enemy has an armour value). Also, outgoing damage is also affected by equipment, so if the player has some gear equipped they will notice that their attacks deal more damage.

In the future, I plan on making the arrows instantly add to your arrow count when picked up instead of having them as a use-able item in the inventory.