Version 0.26
I created a simple menu UI for the game - this allows the user to start on a main menu and view either class select, settings or exit the game. On class select, the player can choose a class and then start the game.

At the moment, this is all UI - the next step is to create the code to choose what weapon the player will start with depending on what class they picked.