Version 0.28
I created some abilities which are based off a Scriptable Object framework which allows me to easily add more. Currently, the player can either dash or use a fireball - these both cost mana and the fireball does 40 base damage (can be higher depending on armour stat boosts).

I need to make a way for players to swap abilities next - however, this should be simple as the scriptable object system already allows for the active ability to be swapped out whenever I want to change it (even while the game is running).