Version 0.32
Fixed the bug - it was due to the position of the sword set in the "SwordIdle" animation - I set this to 0 and it fixed the spawn location of the sword.

I have added a new sword - "Starting Sword". Made previous starting katana to "Bloodsteel Katana". Item swapping works fine with these 2 models which is awesome.

Added framework for different loot pools based on what class you picked. Currently, only warrior loot pool is set up however, but the framework works.

Fixed some code in EquipmentManager to now apply the damage variable of the starting sword - before it was just doing a base value of 50 and would only change the damage value when a new weapon was equipped. Now, it changes the damage value when the starting weapon is equipped. However, I want to change how the damage works and make it so there is no base damage - each weapon will have a damage value and that will be used.