Version 0.36
FIXED THE BUG!!!!! Was due to something in the CheckNeighbour offset - I changed this and also the size of the dungeon from 31x31 to 15x15 as I implemented a system to change direction of generation if the same direction is picked twice (for example, if 2 rooms are generated in the north direction and it tries to generate a third, it will delete the north neighbour from the list and pick either east or west).

I plan on creating bigger prefab rooms, but will keep them all the same size for simplicity with the offset.