Version 0.39
Added health/mana potion UI and made it so you can press V or B to use health or mana pots respectively. Potions are now stored as an integer instead of an item in the inventory.

Changed arrow UI slightly and how arrow restoration works - it goes through the Inventory now instead of being used as an item in the inventory.

Fixed bow/staff rotation bugs.

Currently I have a bug where the model arrow in the bow despawns after a few seconds - I aim to fix this in the next version.