Version 0.41
I created two new rooms - a combat room and a boss teleporter room. The teleporter room is temporary currently (just a room with a red cube in it) and will teleport the player to the boss arena in the future. The combat room is finished - I have added points for enemies to spawn and just need to implement the code to spawn enemies in the rooms. I need to work on implementing a navmesh for each prefab too which I am researching how to do.