Version 0.42
Created 2 more combat rooms and a loot room. These all can generate in the dungeon as well.

If the same room prefab tries to generate twice in a row, it picks another prefab so the player will never go through the same room twice in a row.

I spawned enemies in the rooms successfully and it is based off a 50% chance per enemy (e.g. a room with 10 spawns would, on average, spawn 5 enemies).

I also updated my Unity version from 2021.3.31f1 to 2022.3.23f1 as this new version has an AI package which will allow me to bake a navmesh at runtime so it can create a navmesh for the generated dungeon.

I plan on creating at least one other enemy type in the next version.