Version 0.44
Created enemy arrow and magic projectile prefabs and got them to be created when an enemy of that type attacks. The arrow rotation is bugged currently due to it being set to the rotation of a game object which isnt updated (i am guessing as i had this happen before).

Projectiles also collide with each other which I am going to fix in the next version. Once I get this enemy AI working, I can work on either adding models or creating a boss enemy.