Version 0.45
Range and mage enemies are completed - their projectiles do not collide with each other (including player projectiles and other enemies) and they damage the player.

Added a fireball impact particle as I forgot to before.

There is a bug where the occasional arrow is rotated for some reason - not sure why. If I have the time, I will try to fix it.

Going to work on importing models/animations for the enemies, then getting them to spawn in the dungeon based on enemy type.