Version 0.46
So. Many. Bugs.

Created enemy prefabs - some with a skeleton model and some with capsules for testing. Model kind of works - fire points need to be fixed or I might need to implement a raycast for enemies shooting at the player.

I got the navmesh building at run time and random enemy types spawning in the dungeon, but for some reason they do not move on the nav mesh. I think this may be because the floor does not have a mesh collider currently, so hopefully not a big bug.

Deleted an asset pack as when I tried to animate it, it just did not want to work.