Version 0.48
Created a simple loot bag prefab. Changed roofs of prefab rooms to fix lighting issue. Created a boss arena and a way to teleport the player to it at the end of the dungeon.

There are some bugs with the enemy model and its line of sight on the player/view range - no clue what is causing it but I don't think I have time to fix it.

Need to create a boss next.