Version 0.50
Added more death screen UI - player can restart or go back to main menu.

Replaced current weapon UI with an ability cooldown UI - it displays the current ability and then a green background if it can be used or a red background if it is on cooldown.

Added a new font into the game which suits the medieval theme.

Deleted a lot of assets I am not using.

Quickly created some new images for abilities and weapons.

I plan on adding a "perk" system next version which will let the player pick a buff from 3 random available buffs which they will gain for their playthrough. I will let the player pick this from the main menu and then whenever they complete a dungeon floor.