Version 0.52
Implemented perk system - player can pick a perk after beating a dungeon floor. Does not fully work yet - there is a bug with outlining the UI and actually applying the perk currently, but most of the framework is there.

Currently fixing a bug where the player teleport to the boss room is being overwritten by movement code and not sending the player to the boss room.