Version 0.53
Game is pretty much finished aside from audio, an in-game pause menu and a working settings menu in the main menu.

In this version, I made a way to choose an ability through the main menu - currently there are 2 abilities - a dash and fireball.

A perk can now be picked from the main menu too.

Stats restore on floor completion.